Revert "Merge branch 'doctor-service' of https://github.com/sahanDlenora/hospital-appointment-microservices into doctor-service"
This reverts commit d3c795d475e991771120f38bd64d6df52f8f3815, reversing
changes made to 00fcac1ffd8db39e0b815ba88c8fb0c29a10d81d.

+++ BACKEND/schedule_service/src/main/java/com/hospital/schedule_service/model/Schedule.java
@@ -18,9 +18,9 @@ import java.time.LocalTime;
 public class Schedule {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
-    private Long doctorId;
+    private Integer doctorId;
     private LocalDate date;
     private LocalTime startTime;
     private LocalTime endTime;
